+++ packages/side-effects/side-effect-scroll/types/index.d.ts
@@ -1,3 +1,3 @@
-import { SideEffect } from '@curi/core';
-declare function createScrollSideEffect(): SideEffect;
+import { Subscriber } from '@curi/core';
+declare function createScrollSideEffect(): Subscriber;
 export default createScrollSideEffect;