+++ packages/side-effects/side-effect-title/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Return a `Subscriber`, not a `SideEffect`.
+
 ## 1.0.0-beta.4
 
 * Switch to TypeScript